delete DataSampler
